pyproject.toml
[project]
name = "gemini-calendar-bridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.25.2",
    "boto3>=1.42.59",
    "google-api-python-client>=2.190.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "google-genai>=1.65.0",
    "msgraph-core>=1.3.8",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
